Presentation Version

+++ app/assets/javascripts/views/users/user_listings.js
@@ -6,11 +6,12 @@ Yumster.Views.UserListings = Backbone.CompositeView.extend({
 		this.user = options.user;
 		this.listenTo(this.user, "sync add remove reset", this.render); //logged in user
     this.listenTo(this.model, "sync remove", this.render); //user whose profile is being viewed
-		this.listenTo(this.model.host_meals(), "sync add remove reset", this.render);	
+		this.listenTo(this.model.host_meals(), "sync add remove reset", this.render);
+		// this.listenTo(Yumster.Collections.meals, "sync add remove reset", this.render);	
 	},
 	
 	events: {
-		"click .update": "update",
+		"change .update": "update",
 		"click .meal-click": "viewMeal",
 		"click .modal-button": "newListing"
 	},
@@ -29,6 +30,7 @@ Yumster.Views.UserListings = Backbone.CompositeView.extend({
 		var mealId = $(event.currentTarget).data("meal-id");
 		var mealIndex = $(event.currentTarget).data("meal-ind");
 		var action = $(event.currentTarget).val();
+		$(event.currentTarget).val("");
 		var meal = Yumster.Collections.meals.getOrFetch(mealId);
 		var that = this;
 		if (action === "update") {
@@ -36,7 +38,7 @@ Yumster.Views.UserListings = Backbone.CompositeView.extend({
 		    new Yumster.Views.MealModal({ model: meal, user: this.model });
 		  $('body').prepend(this.modalView.render().$el);
 		  this.modalView.delegateEvents();
-		} else {
+		} else if (action === "delete") {
 		  meal.destroy( {
 		  	success: function () {
 							var listings = that.model.host_meals();
@@ -60,10 +62,11 @@ Yumster.Views.UserListings = Backbone.CompositeView.extend({
 				model: meal,
 				user: this.model,
 				collection: Yumster.Collections.meals,
+				host_meals: this.model.host_meals()
 			});
 		$('body').prepend(this.modalView.render().$el);
 		this.modalView.delegateEvents();
-		setTimeout(this.render, 3000);
+		// setTimeout(this.render.bind(this), 3000);
 	}
 	
 });
